create volumes in an organized directory structure
Feature adds the Controller Publish/Unpublish capability to the driver.

The root directory is already configurable, so that can be matched to
a specific export on the moosefs cluster. We leverage the ability to
set default storage class, trash time, and other settings on that
directory to apply to all cluster volumes.

Within that root directory, a `volumes` directory will contain all of
the volumes created for that cluster.  Later, we will add a snapshots
directory to hold those snapshots, alongside created volumes.