track volume creation and deletion
Moosefs does not have a built-in feature for creating and deleting
volumes, as its goal is to expose a singular filesystem.

This functionality needs persistent storage on the controller, so we
start by mounting the cluster's root directory in the controller.
From there, we can begin to set a directory and file structure.  The
first on this structure is the database file that is used to track the
volumes themselves.